add the correct test file
